♻️ Header에서 MyUtils 분리
