+++ src/components/Header/Header.jsx
@@ -2,35 +2,14 @@ import React from 'react';
 import styled from 'styled-components';
 import Logo from 'components/Header/Logo';
 import GNBWrap from 'components/Header/GNB';
-import IconButton from 'components/common/IconButton';
-
-const iconButtons = {
-  search: {
-    width: '24px',
-    height: '24px'
-  },
-  user: {
-    width: '22px',
-    height: '22px'
-  },
-  cart: {
-    width: '27px',
-    height: '28px'
-  }
-};
+import MyUtils from 'components/Header/MyUtils';
 
 export default function Header() {
   return (
     <Wrap>
       <Logo size="2.857rem" />
       <GNBWrap />
-      <MyUtils>
-        {Object.keys(iconButtons).map((iconName, index) => (
-          <li key={`${iconName}_${index}`}>
-            <IconButton icon={iconName} width={iconButtons[iconName].width} height={iconButtons[iconName].hieght} />
-          </li>
-        ))}
-      </MyUtils>
+      <MyUtils />
     </Wrap>
   );
 }
@@ -45,9 +24,4 @@ const Wrap = styled.header({
   gap: '40px',
   padding: '30px 80px 16px',
   borderBottom: `1px solid ${({ theme }) => theme.COLOR.BLACK[100]}`
-});
-
-const MyUtils = styled.ul({
-  display: 'flex',
-  gap: '15px'
 });
